Default to fixed mass resolution rather than using a relative resolution
Had seen indications that this was important for getting converged results at z=1. In hindsight I suspect this is probably because the fractional resolution was being set relative to the z=0 halo mass, so with 0.001 fractional mass resolution, a 10^12 halo at z=1 that will become a satellite of a 10^14 halo at z=0 is simulated with only 10^11 mass resolution.